Realized there is no need to make the geom object private.
